pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pylate"
dynamic = ["version"]
description = "A library for training and retrieval with ColBERT."
readme = "README.md"
authors = [{ name = "LightOn" }]
license = "MIT"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "sentence-transformers == 5.1.1",
    "datasets >= 2.20.0",
    "accelerate >= 0.31.0",
    "pandas >= 2.2.1",
    "transformers >= 4.41.0, <= 4.56.2",
    "ujson == 5.10.0",
    "ninja == 1.11.1.4",
    "fastkmeans == 0.5.0",
    "fast-plaid>=1.2.4.260,<=1.3.0.290",
]
keywords = []

[tool.setuptools.dynamic]
version = { attr = "pylate.__version__" }

[project.urls]
Homepage = "https://github.com/lightonai/pylate.git"
Documentation = "https://lightonai.github.io/pylate/"

[project.optional-dependencies]
dev = [
    "ruff >= 0.4.9",
    "pytest-cov >= 5.0.0",
    "pytest-xdist >=3.6.0",
    "pytest-rerunfailures >= 15.0.0",
    "pytest >= 8.2.1",
    "pandas >= 2.2.1",
    "einops>=0.8.1",
    "pre-commit>=4.1.0",
    "ranx >= 0.3.16",
    "beir >= 2.0.0",
    "fastapi >= 0.114.1",
    "uvicorn >= 0.30.6",
    "batched >= 0.1.2",
    "voyager >= 2.0.9",
    "typos >= 0.11.0",
]
eval = ["ranx >= 0.3.16", "beir >= 2.0.0"]
api = ["fastapi >= 0.114.1", "uvicorn >= 0.30.6", "batched >= 0.1.2"]
voyager = ["voyager >= 2.0.9"]
scann = ["scann >= 1.4.2"]


[dependency-groups]
docs = [
    "mkdocs-material == 9.5.32",
    "mkdocs-awesome-pages-plugin == 2.9.3",
    "mkdocs-jupyter == 0.24.8",
    "mkdocs_charts_plugin == 0.0.10",
    "numpydoc == 1.8.0",
]

[tool.setuptools]
packages = ["pylate"]

[tool.setuptools.package-data]
pylate = ["hf_hub/model_card_template.md", "py.typed"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "D417"]
ignore = ["ANN"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::RuntimeWarning",
    "ignore::UserWarning",
]
addopts = "--doctest-modules --verbose -ra --cov-config=.coveragerc -m \"not web and not slow\""
doctest_optionflags = "NORMALIZE_WHITESPACE NUMBER"
norecursedirs = ["build", "docs", "node_modules", "pylate/server"]
markers = ["model_loading: marks tests that download large models"]

[tool.typos.default]
binary = false
locale = "en"
extend-ignore-words-re = [
    "arange",
]
extend-ignore-re = [
    "xlm-mlm-ende-1024",
    "xlm-clm-ende-1024",
    "compute_metrices",
]
